Time: 7 ms (44.31%), Space: 19 MB (30.75%), Date: 02-14-2026

+++ LeetCode/Medium/0799-champagne-tower/0799-champagne-tower.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     double champagneTower(int poured, int query_row, int query_glass) {
-        if (query_row == 0) return min(poured, 1); 
+        if (query_row == 0) return min((double)poured, 1.0); 
 
         vector<double> dp(102, 0.0);
-        dp[1] = poured;
+        dp[1] = (double)poured;
         int i = 1;
         while (i <= query_row) {
             vector<double> _dp(102, 0.0);
@@ -15,7 +15,6 @@ public:
             dp = _dp;
 
             if (i == query_row) return min(dp[query_glass + 1], 1.0);
-
             i++;
         }
